# Conflicts:
#	community/source/insert.php

+++ shop_notice/source/view.php
@@ -12,6 +12,13 @@ if(isset($_GET['no'])){
 include_once '../../common_lib/createLink_db.php';
 include_once '../../shopping_lib/create_table_notice.php';
 
+$sql= "select * from shop_notice";
+$result= mysqli_query($con, $sql);
+$row2=mysqli_num_rows($result);
+
+$avant=$no-1;
+$next=$no+1;
+
 $sql= "select * from shop_notice where notice_no=$no";
 $result= mysqli_query($con, $sql);
 $row=mysqli_fetch_array($result);
@@ -29,7 +36,7 @@ $file_copied[1]=$row['file_copied_1'];
 <head>
   <meta charset="utf-8">
   <title>야! 몰</title>
-  <link rel="stylesheet" href="../css/notice.css?ver=2">
+  <link rel="stylesheet" href="../css/notice.css?ver=4">
   <link rel="stylesheet" href="../../shopping/css/cart.css?ver=1">
   <link rel="stylesheet" href="../../common_css/shop_index_css3.css">
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
@@ -78,10 +85,44 @@ $file_copied[1]=$row['file_copied_1'];
     			<tr id="notice_table4_tr">
     				<td id="notice_table4_tr_td">
     					<a href="./shop_notice.php?page=1"><input type="button" value="목록" class="notice_view_button"></a>
+				        <?php 
+                        if(isset($id) && ($id==="admin")){
+                        ?>
     					<a href="./notice_write_form.php?no=<?=$no?>&mode=modify"><input type="button" value="수정" class="notice_view_button"></a>
     					<a href="./notice_delete.php?no=<?=$no?>"><input type="button" value="삭제" class="notice_view_button"></a>
+    					<?php
+                        }
+                        ?>
     				</td>
     			</tr>
+    			<?php 
+    			if($no != 1){
+    			    $sql= "select * from shop_notice where notice_no=$avant";
+    			    $result= mysqli_query($con, $sql);
+    			    $row=mysqli_fetch_array($result);
+    			    
+    			    $nick=$row['notice_nick'];
+    			    $subject=$row['notice_subject'];
+    			?>
+    			<tr id="avant_list_tr">
+    				<td id="avant_list_td">▲&nbsp;이전글&nbsp;&nbsp;&nbsp;&nbsp;<a href="./view.php?no=<?=$avant?>"><?=$subject?></a></td>
+    			</tr>
+    			<?php 
+    			}
+    			if($row2 != $no){	
+    			    $sql= "select * from shop_notice where notice_no=$next";
+    			    $result= mysqli_query($con, $sql);
+    			    $row=mysqli_fetch_array($result);
+    			    
+    			    $nick=$row['notice_nick'];
+    			    $subject=$row['notice_subject'];
+                ?>
+                <tr id="next_list_tr">
+                    <td id="next_list_td">▼&nbsp;다음글&nbsp;&nbsp;&nbsp;&nbsp;<a href="./view.php?no=<?=$next?>"><?=$subject?></a></td>   
+				</tr>
+                <?php 
+    			}
+                ?>	
     		</table>
     	</div>
     </section>
